==> Machine_18/notebooks/06_merge_raw_jobs_and_clean_jobevents_dataset.ipynb <==
# 06 - Création du dataset des données de jobs et de job_events fusionnées

Ce notebook génère 1 csv :

- merge_raw_jobs_and_clean_jobevents_dataset.csv

Etapes :

- Suppression des colonnes doublon dans les 2 datasets

- Fusion des datasets

# A. Imports

## a) Librairies

import os, json
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

## b) Données

# source path
jobs = '../data/jobs/raw_jobs_dataset.csv'

# source path
events = '../data/jobs/clean_merge_job_events_dataset.csv'

# save target path
save_csv = '../data/jobs/merge_raw_jobs_and_clean_jobevents_dataset.csv'

# B. Dataframe

## a) Création des dataframes jobs et events

### 1. jobs

# création d'un dataframe à partir du csv de données
jobs_df = pd.read_csv(jobs, index_col=0)
jobs_df.info()

### 2. job events

# création d'un dataframe à partir du csv de données
events_df = pd.read_csv(events, index_col=0)
events_df.info()

## b) Création d'un dataframe fusionné

### 1. Suppression de colonnes

# suppression des colonnes avec des valeurs uniques
for col in list(jobs_df.columns):
    if jobs_df[col].nunique() == 1:
        jobs_df.drop(col, axis=1, inplace=True)

# suppression des colonnes avec des valeurs uniques
for col in list(events_df.columns) :
    if events_df[col].nunique() == 1:
        events_df.drop(col, axis=1, inplace=True)

### 2. Concordance des colonnes entre dataframes

# on renomme des colonnes de jobs
jobs_df = jobs_df.rename(columns={
 'uses_ifoil' : 'ifoil',
 'iper_bvar_count': 'bars'
 })

# on renomme des colonnes de events
events_df = events_df.rename(columns={
    'totalCopies_start':'total_copies_requested',
    'LED_iper' : 'LED',
    'bars_iper' : 'bars',
    'drops_iper': 'drops',
    'dithering_iper' : 'dithering',
    'deadPixelsOffset_iper' : 'deadPixelsOffset',
    'level_user' : 'operator_level',
    'operator_user':'operator',
    'enabled_ifoil' : 'ifoil',
    'x_imageLayout_layout' : 'x_imageLayout',
    'y_imageLayout_layout' : 'y_imageLayout',
    'name_paperFormat_layout' : 'paperName',
    'width_paperFormat_layout' : 'paperWidth',
    'height_paperFormat_layout' : 'paperHeight',
    'speed_layout' : 'speed',
    'topMargin_registration_remoteScannerRegistration' : 'topMargin_remoteScannerRegistration',
    'leftMargin_registration_remoteScannerRegistration' : 'leftMargin_remoteScannerRegistration',
    'totalCopies_end' : 'total_copies',
    'consumption_operatorSideTanks_varnishConsumption' : 'varnishConsumptionVarnish_3d'
    })

# Comparer les valeurs de la colonne "bars" entre les dataframes "events_df" et "jobs_df"
events_df['bars'].isin(jobs_df['bars']).any()

# liste des colonnes communes
common_cols = [col for col in jobs_df.columns.to_list() if col in events_df.columns.to_list()]

### 3. Concordance des valeurs entre colonnes communes

jobs_df.shape

events_df.shape

events_df.columns

# Colonnes de events_df_17
cols_events_df_17 = ['jobId', 'timestamp_start', 'total_copies_requested', 'LED', 'bars',
                     'drops', 'dithering', 'deadPixelsOffset', 'operator_level', 'operator',
                     'speed_ifoil', 'ifoil', 'optifoil_ifoil', 'stampAreas_ifoil',
                     'heater1Enabled_ifoil', 'speedTensionIn_ifoil', 'speedTensionOut_ifoil',
                     'heater1Temperature_ifoil', 'x_imageLayout', 'y_imageLayout',
                     'paperName', 'paperWidth', 'paperHeight', 'speed', 'power_irDryers',
                     'power_uvDryers', 'redScore_gridMode_remoteScannerRegistration',
                     'redScore_cropmarksMode_remoteScannerRegistration',
                     'x_cropmark1_cropmarksMode_remoteScannerRegistration',
                     'y_cropmark1_cropmarksMode_remoteScannerRegistration',
                     'x_cropmark2_cropmarksMode_remoteScannerRegistration',
                     'y_cropmark2_cropmarksMode_remoteScannerRegistration',
                     'exposureTime_manualLighting_remoteScannerRegistration',
                     'redScore_fullScannerMode_remoteScannerRegistration',
                     'blueScore_fullScannerMode_remoteScannerRegistration',
                     'greenScore_fullScannerMode_remoteScannerRegistration',
                     'enable_specialSubstrate_remoteScannerRegistration',
                     'mode_remoteScannerRegistration', 'jobState', 'timestamp_end',
                     'total_copies', 'varnishConsumptionVarnish_3d']

# Colonnes de events_df_14
cols_events_df_14 = ['jobId', 'timestamp_start', 'total_copies_requested', 'LED', 'bars',
                     'drops', 'dithering', 'deadPixelsOffset', 'operator_level', 'operator',
                     'speed_ifoil', 'ifoil', 'optifoil_ifoil', 'stampAreas_ifoil',
                     'heater1Enabled_ifoil', 'speedTensionIn_ifoil',
                     'heater1Temperature_ifoil', 'x_imageLayout', 'y_imageLayout',
                     'paperName', 'paperWidth', 'paperHeight', 'speed', 'power_irDryers',
                     'power_uvDryers', 'topMargin_remoteScannerRegistration',
                     'leftMargin_remoteScannerRegistration',
                     'redScore_gridMode_remoteScannerRegistration',
                     'redScore_cropmarksMode_remoteScannerRegistration',
                     'redScore_fullScannerMode_remoteScannerRegistration',
                     'blueScore_fullScannerMode_remoteScannerRegistration',
                     'greenScore_fullScannerMode_remoteScannerRegistration',
                     'mode_remoteScannerRegistration', 'jobState', 'timestamp_end',
                     'total_copies', 'varnishConsumptionVarnish_3d']

# Trouver les colonnes uniques à events_df_17
unique_cols_to_df_17 = set(cols_events_df_17) - set(cols_events_df_14)

print(unique_cols_to_df_17)

jobs_df.head()

events_df.head()

events_df.shape

jobs_df.shape

# on supprime les lignes dans events_df qui n'ont pas de JobId qui coïncident avec 'id_on_machine' de jobs_df
events_df = events_df[events_df['jobId'].isin(jobs_df['id_on_machine'])]

events_df.shape

# # on liste les colonnes des deux dataframes contenant toutes leurs valeurs identiques
# col_with_duplicates = []
# for col in common_cols :
#     if (events_df[col].values == jobs_df[col].values).all():
#         col_with_duplicates.append(col)
# col_with_duplicates
col_with_duplicates = []
for col in common_cols:
    # Vérifier si les deux colonnes ont la même longueur
    if len(events_df[col]) == len(jobs_df[col]):
        # Si elles ont la même longueur, effectuer la comparaison
        if (events_df[col].values == jobs_df[col].values).all():
            col_with_duplicates.append(col)
    else:
        print(f"Les colonnes {col} n'ont pas la même longueur.")
col_with_duplicates

# suppression des colonnes de jobs en doublons dans events
jobs_df = jobs_df.drop(col_with_duplicates, axis=1)

### 4. Fusion des dataframes

merge_df = pd.merge(jobs_df, events_df, left_on='id_on_machine', right_on='jobId', suffixes=['_job', '_event'])
merge_df = merge_df.drop('id_on_machine', axis=1)

### 5. Vérification

# on verifie que les jobId sont uniques
merge_df.jobId.duplicated().any()

merge_df.head()

merge_df.columns

# id temoin
check_jobid = 1645522984
# détail de la ligne témoin dans chaque dataset
jobs_check_line = jobs_df[jobs_df.id_on_machine == check_jobid]
events_check_line = events_df[events_df.jobId == check_jobid]
merge_check_line = merge_df[merge_df.jobId == check_jobid]

jobs_check_line

events_check_line

merge_check_line

# on conserve les colonnes started_at et ended_at
for index, row in merge_df.iterrows():
    # on verifie que started_at à le datetime le plus petit
    if row.timestamp_start < row.started_at :
        row.started_at = row.timestamp_start
    # on verifie que ended_at à le datetime le plus grand
    if row.timestamp_end > row.ended_at :
        row.ended_at = row.timestamp_end

merge_df = merge_df.drop(['timestamp_start','timestamp_end'], axis=1)

# # conversion des colonnes contenant des valeurs de temsps au format datetime
# jobs_df['started_at'] = pd.to_datetime(jobs_df['started_at'], utc=True)
# jobs_df['ended_at'] = pd.to_datetime(jobs_df['ended_at'], utc=True)
# events_df['timestamp_start'] = pd.to_datetime(events_df['timestamp_start'], utc=True)
# events_df['timestamp_end'] = pd.to_datetime(events_df['timestamp_end'], utc=True)

### 6. Output csv

# sauvegarde du dataset en csv
merge_df.to_csv(save_csv)

merge_df.head(3)

for col in merge_df.columns:
    unique_values = merge_df[col].nunique()
    print(f"Column {col} has {unique_values} unique values.")


merge_df.info()

# C. Visualisation

df = merge_df.copy()

df.jobState.unique()

df.columns

# Colonnes de events_df_17
df17 = ['total_copies_job', 'started_at', 'ended_at', 'paperHeight_job',
       'paperWidth_job', 'scanner_mode', 'bars_job',
       'varnishConsumptionVarnish_3d_job', 'run', 'total_run',
       'copies_per_run', 'jobId', 'total_copies_requested', 'LED',
       'bars_event', 'drops', 'dithering', 'deadPixelsOffset',
       'operator_level', 'operator', 'speed_ifoil', 'ifoil', 'optifoil_ifoil',
       'stampAreas_ifoil', 'heater1Enabled_ifoil', 'speedTensionIn_ifoil',
       'speedTensionOut_ifoil', 'heater1Temperature_ifoil', 'x_imageLayout',
       'y_imageLayout', 'paperName', 'paperWidth_event', 'paperHeight_event',
       'speed', 'power_irDryers', 'power_uvDryers',
       'redScore_gridMode_remoteScannerRegistration',
       'redScore_cropmarksMode_remoteScannerRegistration',
       'x_cropmark1_cropmarksMode_remoteScannerRegistration',
       'y_cropmark1_cropmarksMode_remoteScannerRegistration',
       'x_cropmark2_cropmarksMode_remoteScannerRegistration',
       'y_cropmark2_cropmarksMode_remoteScannerRegistration',
       'exposureTime_manualLighting_remoteScannerRegistration',
       'redScore_fullScannerMode_remoteScannerRegistration',
       'blueScore_fullScannerMode_remoteScannerRegistration',
       'greenScore_fullScannerMode_remoteScannerRegistration',
       'enable_specialSubstrate_remoteScannerRegistration',
       'mode_remoteScannerRegistration', 'jobState', 'total_copies_event',
       'varnishConsumptionVarnish_3d_event']

# Colonnes de events_df_18
df18 = ['total_copies_job', 'started_at', 'ended_at', 'speed_job',
       'paperHeight_job', 'paperWidth_job', 'total_copies_requested_job',
       'ifoil_job', 'scanner_mode', 'bars_job',
       'varnishConsumptionVarnish_3d_job', 'run', 'total_run',
       'copies_per_run', 'jobId', 'total_copies_requested_event', 'LED',
       'bars_event', 'drops', 'dithering', 'deadPixelsOffset',
       'operator_level', 'operator', 'speed_ifoil', 'ifoil_event',
       'optifoil_ifoil', 'vacuumIn_ifoil', 'vacuumOut_ifoil',
       'stampAreas_ifoil', 'heater1Enabled_ifoil', 'speedTensionIn_ifoil',
       'speedTensionOut_ifoil', 'heater1Temperature_ifoil', 'x_imageLayout',
       'y_imageLayout', 'paperName', 'paperWidth_event', 'paperHeight_event',
       'speed_event', 'power_irDryers', 'power_uvDryers',
       'redScore_gridMode_remoteScannerRegistration',
       'redScore_cropmarksMode_remoteScannerRegistration',
       'x_cropmark1_cropmarksMode_remoteScannerRegistration',
       'y_cropmark1_cropmarksMode_remoteScannerRegistration',
       'x_cropmark2_cropmarksMode_remoteScannerRegistration',
       'y_cropmark2_cropmarksMode_remoteScannerRegistration',
       'exposureTime_manualLighting_remoteScannerRegistration',
       'redScore_fullScannerMode_remoteScannerRegistration',
       'blueScore_fullScannerMode_remoteScannerRegistration',
       'greenScore_fullScannerMode_remoteScannerRegistration',
       'enable_specialSubstrate_remoteScannerRegistration',
       'mode_remoteScannerRegistration', 'jobState', 'total_copies_event',
       'varnishConsumptionVarnish_3d_event']

# df14 et df17 sont vos listes de colonnes

# Convertir les listes en ensembles
colonnes_df17 = set(df17)
colonnes_df18 = set(df18)

# Trouver les colonnes uniques à df17
colonnes_uniques_df17 = colonnes_df17 - colonnes_df18
print("Colonnes uniques à df17 (machine 17) :")
print(colonnes_uniques_df17)

# Trouver les colonnes uniques à df14
colonnes_uniques_df18 = colonnes_df18 - colonnes_df17
print("\nColonnes uniques à df18 (machine 18) :")
print(colonnes_uniques_df18)
